apps/api/requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
psycopg2-binary
python-dotenv
alembic
pydantic
requests